Move export commands to own file
